Fixing the logger extensions to take in params object[] instead of string[]
Fixes bug # https://github.com/aspnet/Logging/issues/89

Right now the logger extensions take in params string[] as args for substitution values. This makes it painful to send basic CLR types and user code has to do a .ToString() on even basic CLR types.

With this change it will be possible to possible the object as it is.