Adding some tests for the MapPath extension method
And looks like I pushed the MapPath extension change too soon. I had to fix an issue in the path for forward and backward slash.

For example if an application tries to do MapPath("/resource") then this code does not work cross plat unless the directory separator characters are fixed appropriately.

Fixed it and also added some test coverage.