Triggering TestHost with a message
This change will allow TestHost to execute a command (discover or execute)
based on a message instead of parsing the commandline. This will get us
out of trouble for the cases where a 'run all tests' results in a
commandline string that is too long.

Updates the UI and test code as well to use this.