Add support for proxying IList<>
This change adds the ability to generate a proxy type which can expose an
IList<T> as an IReadOnlyList<U> - where the conversion from T -> is via a
second proxy.

We may need to support for additional collection types (IEnumerable,
IDictionary) in the future. This change should get the codebase to a point
where we're able to start using it.