Stop producing Microsoft.Extensions.CommandLineUtils
This library was only intended for usage with ASP.NET Core's tools, and not a general purpose library. After receiving bugs and questions about this, we realized that we didn't really intend to produce and support a command-line parsing library.

We will still leave the 1.0.0 and 1.1.0 versions of this library on NuGet.org, but the current plan is to stop producing new versions of this package.

The source code is still available for internal use via Microsoft.Extensions.CommandLineUtils.Sources, however we will not be publishing this library again as a general purpose library.