Add named clients and options and caching
Reflects the chosen naming of IHttpClientFactory and System.Net.Http
namespace.

Adds the ability to create named clients
- built on top of named options
- this means introducing an options type

Also adds caching of handlers

Resolves: #11 #4 #5 #6

----------------

The options type looks a little wonky but humour me for a second here.

Fundamentally I think it's the right strategy for us to cache *handlers*
not clients. The handler is the thing that actually does connection
management, and conveniently they are opaque and immutable once
constructed.

The HttpClient on the other hand *is* mutable. So if you do something
like dispose it, or modify the big public mutable things, you do so for
the rest of the callers in the system.

I think that caching handlers, but not clients gets us everything we
want. In addition using the HttpClient with either pattern (keep forever
or dispose) will 'just work' without any complicated code or drawbacks
for users.

------

So after establishing that, we need to a way to express both deferred
operations that build a 'handler' pipeline, and deferred operations that
modify an HttpClient - and that's what it looks like on the options
type.

We should add extension methods or maybe another builder if necessary to
smooth it over, but I think we'll get pretty far with what I have until
we start tying in 3rd party libraries.