Implementation of typed clients (#31)
aspnet/HttpClientFactory#17

This PR has **infrastructure** for typed clients. You either need to
write your own client classes, or use a library like refit.

What's here is the general pattern and plumbing. The intent is that a
library like refit would implement ITypedHttpClientFactory, OR you would
plug it using the overloads of AddTypedClient<T> that take a factory
function.