[Design] Add handler rotation and expiry
This adds a simple approach to handler rotation and handler expiry that
should have virtually no impact on how users consume HttpClient.

The idea is that handlers sit in an 'active' pool until they expire
(with a configuable fixed lifetime). While they are in the 'active' pool
they can be shared amongst as many HttpClient instances as we need to
create.

When a handler that supports expiry is created, a timer will be queued
to atomically remove it from the pool. A new one will then be created to
take its place.

----

Ultimately we need to dispose the handler that we removed from the
active pool, but it's not safe to do that right away because there are
probably existing leases. Ultimately this is a task for the GC.

When we create the handler, the one that we give to HttpClients is an
instance of a type we control, that is safe to GC without the need for
disposal.

Once we move the handler to the 'expiry' pool, we can downgrade our
reference to the tracking handler to a weak reference. Once we see that
its GCed, it's safe to dispose the inner handler.

----

We ultimately process all of these disposal jobs on a fixed timer that
processes all of the 'expired' items.

Since all of the data flow is unidirectional there isn't much threading
heavy-lifting.