Fix #124 and #134
This is a rework of the http client factory interacts with dependency
injection when creating a message handler. Since the http client factory
manages the lifetimes of message handlers, it's not appropriate for a
message handler and related services to share lifetimes with the
application or 'current request' scope.

Now the factory creates a scope each time it creates a handler, and then
disposes the scope when the handler is disposed. This allows the handler
building process to resolve scoped services and also to prevent
allocating lots and lots of disposable transients in the application
global scope.

The fix for #134 also required a rework of some details of
`DefaultTypedClientFactory<>`.