Allow duplicate class to AddHttpClient<T>
Fixes: #2077

In 3.0 we introduced validation to try and prevent some cases of invalid
usage on the factory that had lead to user bug reports.

Unfortunately we blocked a few legitimate usage scenarios behind
expections.

In this case the common usage is for a library to register a typed
client with `AddHttpClient<MyClient>(...)`. User code can then
collaborate by calling the same thing, and interacting with the builder
that's returned.

This change explicitly allows this pattern by fine-tuning the
validation.